OOP Modifications : token edits are now deriving from the same object
