+++ manage.py
@@ -73,5 +73,6 @@ def run():
     """
     app.run()
 
+
 if __name__ == '__main__':
     manager.run()
