+++ tests/test_selenium/test_token_similar_edit.py
@@ -1,0 +1,7 @@
+from tests.test_selenium.base import TokenEdit2CorporaBase
+
+
+class TokenEditBase(TokenEdit2CorporaBase):
+    """ Base class with helpers to test token edition page """
+    CORPUS = "wauchier"
+    CORPUS_ID = "1"
